+++ src/components/Detail/useDetailNavigation.tsx
@@ -4,7 +4,7 @@ import { SearchResult } from "../../model/Search.ts";
 import { useSearchStore } from "../../stores/search/search-store.ts";
 import { LAST_SEARCH_RESULT } from "../Search/SearchUrlParams.ts";
 import { isNumber, isString } from "lodash";
-import { detailTier2Path } from "../Text/Annotated/project/utils/detailPath.ts";
+import { detailTier2Path } from "../../utils/detailPath.ts";
 import { useUrlSearchParamsStore } from "../Search/useSearchUrlParamsStore.ts";
 
 export type DetailTierAndParams = {
