--- src/components/Text/Annotated/project/EntitySummaryProps.ts
@@ -1,3 +1,0 @@
-import { AnnoRepoBody } from "../../../../model/AnnoRepoAnnotation.ts";
-
-export type EntitySummaryProps = { body: AnnoRepoBody };
