+++ src/components/Text/Annotated/utils/getAnnotationsByTypes.ts
@@ -1,4 +1,4 @@
-import { AnnoRepoAnnotation } from "../../../../../model/AnnoRepoAnnotation.ts";
+import { AnnoRepoAnnotation } from "../../../../model/AnnoRepoAnnotation.ts";
 
 export function getAnnotationsByTypes(
   annotations: AnnoRepoAnnotation[],
