+++ src/model/EntitySummaryProps.ts
@@ -1,0 +1,3 @@
+import { AnnoRepoBody } from "./AnnoRepoAnnotation.ts";
+
+export type EntitySummaryProps = { body: AnnoRepoBody };
