+++ src/projects/brederode/Header.tsx
@@ -5,7 +5,7 @@ import {
   useProjectStore,
 } from "../../stores/project.ts";
 import { matchPath, useLocation, useNavigate } from "react-router";
-import { detailTier2Path } from "../../components/Text/Annotated/project/utils/detailPath.ts";
+import { detailTier2Path } from "../../utils/detailPath.ts";
 import { useAnnotationStore } from "../../stores/annotation.ts";
 import { Button } from "react-aria-components";
 import { isLetterBody } from "../kunstenaarsbrieven/annotation/ProjectAnnotationModel.ts";
