+++ src/utils/detailPath.ts
