+++ src/utils/trimMiddle.ts
